+++ ext/SymbolicsGroebnerExt.jl
@@ -1,0 +1,73 @@
+module SymbolicsGroebnerExt
+
+using Groebner
+
+if isdefined(Base, :get_extension)
+    using Symbolics
+    using Symbolics: Num, symtype
+else
+    using ..Symbolics
+    using ..Symbolics: Num, symtype
+end
+
+"""
+    groebner_basis(polynomials; kwargs...)
+
+Computes a Groebner basis of the ideal generated by the given `polynomials`
+using Groebner.jl as the backend.
+
+The basis is guaranteed to be unique.
+The algorithm is randomized, and the output is correct with high probability.
+
+If a coefficient in the resulting basis becomes too large to be represented
+exactly, `DomainError` is thrown.
+
+## Optional Arguments
+
+The Groebner.jl backend provides a number of useful keyword arguments, which are
+also available for this function. See `?Groebner.groebner`.
+
+## Example
+
+```jldoctest
+julia> using Symbolics, Groebner
+
+julia> @variables x y;
+
+julia> groebner_basis([x*y^2 + x, x^2*y + y])
+```
+"""
+function Symbolics.groebner_basis(polynomials::Vector{Num}; kwargs...)
+    polynoms, pvar2sym, sym2term = Symbolics.symbol_to_poly(polynomials)
+    basis = Groebner.groebner(polynoms; kwargs...)
+    PolyType = symtype(first(polynomials))
+    Symbolics.poly_to_symbol(basis, pvar2sym, sym2term, PolyType)
+end
+
+"""
+    is_groebner_basis(polynomials; kwargs...)
+
+Checks whether the given `polynomials` forms a Groebner basis using Groebner.jl
+as the backend.
+
+## Optional Arguments
+
+The Groebner.jl backend provides a number of useful keyword arguments, which are
+also available for this function. See `?Groebner.isgroebner`.
+
+## Example
+
+```jldoctest
+julia> using Symbolics, Groebner
+
+julia> @variables x y;
+
+julia> is_groebner_basis([x^2 - y^2, x*y^2 + x, y^3 + y])
+```
+"""
+function Symbolics.is_groebner_basis(polynomials::Vector{Num}; kwargs...)
+    polynoms, _, _ = Symbolics.symbol_to_poly(polynomials)
+    Groebner.isgroebner(polynoms; kwargs...)
+end
+
+end # module
